blockchain/requirements.txt
# Blockchain and Polygon integration dependencies
web3>=7.13.0,<8.0.0
eth-account>=0.13.7,<0.14.0
eth-tester>=0.13.0b1,<0.14.0

# Solidity compiler for smart contract deployment
py-solc-x>=2.0.0

# Polygon-specific libraries
polygon-sdk-py>=1.0.0
py-evm>=0.8.0

# Cryptography and hashing
cryptography>=41.0.0
pycryptodome>=3.19.0

# HTTP and networking
requests>=2.31.0
aiohttp>=3.9.0

# Data handling and validation
pydantic>=2.5.0
dataclasses-json>=0.6.0

# Environment and configuration
python-dotenv>=1.0.0

# Testing dependencies (for blockchain testing)
pytest>=7.4.0
pytest-asyncio>=0.21.0